+++ apps/community/src/app/(board)/news/[id]/page.tsx
