+++ apps/community/src/app/(board)/notice/[id]/page.tsx
