+++ apps/community/src/app/(board)/notice/page.css.ts
